+++ lib/src/factorial.dart
@@ -1,37 +1,35 @@
 import 'dart:math' as math;
 
-import 'package:bignum/bignum.dart';
-
 /// Computes factorial of [n].
 ///
 /// Uses the efficient FactorialSplit algorithm:
 /// https://github.com/PeterLuschny/Fast-Factorial-Functions/blob/master/JavaFactorial/src/de/luschny/math/factorial/FactorialSplit.java
-BigInteger factorial(int n) {
+BigInt factorial(int n) {
   if (n < 0) {
     throw new ArgumentError("Factorial of $n is undefined.");
   }
 
-  if (n < 2) return new BigInteger(1);
+  if (n < 2) return BigInt.one;
 
-  BigInteger p = BigInteger.ONE;
-  BigInteger r = BigInteger.ONE;
+  BigInt p = BigInt.one;
+  BigInt r = BigInt.one;
   int N = 1;
 
-  BigInteger product(int n) {
+  BigInt product(int n) {
     final int m = n ~/ 2;
     if (m == 0) {
-      return new BigInteger(N += 2);
+      return new BigInt.from(N += 2);
     }
     if (n == 2) {
-      return new BigInteger((N += 2) * (N += 2));
+      return new BigInt.from((N += 2) * (N += 2));
     }
-    return product(n - m).multiply(product(m));
+    return product(n - m) * product(m);
   }
 
   int h = 0;
   int shift = 0;
   int high = 1;
-  int log2n = (math.log(n) / math.LN2).floor();
+  int log2n = (math.log(n) / math.ln2).floor();
 
   while (h != n) {
     shift += h;