+++ lib/src/t_distribution.dart
@@ -5,8 +5,8 @@ part of t_stats;
 /// Stolen from https://github.com/dart-lang/ton80/blob/master/bin/ton80.dart.
 double _computeTDistribution(int n) {
   const List<double> table = const [
-    double.NAN,
-    double.NAN,
+    double.nan,
+    double.nan,
     12.71,
     4.30,
     3.18,