+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # Changelog
 
+## 2.0.0
+
+- Upgrade to Dart 2
+- Use BigInt (found in `dart:core`) instead of `package:bignum`. This
+  affects return values as well (factorial) so we're bumping major semver
+  of this package to 2.0.
+
 ## 1.1.2
 
 - Upgrade to latest